Long to int

+++ src/Main.java
@@ -41,8 +41,8 @@ public class Main {
         if(n < 2) return false;
         if(n == 2 || n == 3) return true;
         if(n%2 == 0 || n%3 == 0) return false;
-        long sqrtN = (int)Math.sqrt(n)+1;
-        for(long i = 6L; i <= sqrtN; i += 6) {
+        int sqrtN = (int)Math.sqrt(n)+1;
+        for(int i = 6; i <= sqrtN; i += 6) {
             if(n%(i-1) == 0 || n%(i+1) == 0) return false;
         }
         return true;
